fixing and improving handling of AiP RequestFeeRemission event
